+++ tests/LapsTestCase.php
@@ -10,12 +10,12 @@ class LapsTestCase extends TestCase {
 
 	use MockeryPHPUnitIntegration;
 
-	protected function setUp() {
+	protected function setUp():void {
 		parent::setUp();
 		Monkey\setUp();
 	}
 
-	protected function tearDown() {
+	protected function tearDown(): void {
 		Monkey\tearDown();
 		parent::tearDown();
 	}